split plugin into submodule
